% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCacheMatrix.R
\name{makeCacheMatrix}
\alias{makeCacheMatrix}
\title{Returns the list is used to cache the original square matrix and the cached inverse}
\usage{
mat<-makeCacheMatrix(rnorm(9,3,3)) # assuming the matrix generted is invertible
mat$get()  # retrives the input matrix
mat$getInverse() # retrives the inverse of matrix
}
\arguments{
\item{x}{- Square matrix}
}
\value{
A list with input data matrix and placeholder for caching the inverse.
}
\description{
Builds necessary data structure to capture the input ,
and provides necessary infrastructure for caching the inverse of matrix
}
\author{
Priyesh Kannan
}
